refactor: extract A&R office router (PR-7, Phase 1)
